# swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  Email:
    description: メールアドレス (default view)
    example:
      email: yKicchanApp@gmail.com
    properties:
      email:
        description: メールアドレス
        example: yKicchanApp@gmail.com
        format: email
        type: string
    required:
    - email
    title: 'Mediatype identifier: application/vnd.email+json; view=default'
    type: object
...
  File_path:
    description: アップロード済みのファイルのパス (default view)
    example:
      path: /files/image.png
    properties:
      path:
        description: ファイルのパス
        example: /files/image.png
        type: string
    required:
    - path
    title: 'Mediatype identifier: application/vnd.file_path+json; view=default'
    type: object
  File_pathCollection:
    description: File_pathCollection is the media type for an array of File_path (default
      view)
    example:
    - path: /files/image.png
    - path: /files/image.png
    items:
      $ref: '#/definitions/File_path'
    title: 'Mediatype identifier: application/vnd.file_path+json; type=collection;
      view=default'
    type: array
...
  LoginPayload:
    description: 認証時に受け取るログイン情報
    example:
      id: yKicchan
      password: password
    properties:
      id:
        description: ユーザーID
        example: yKicchan
        minLength: 1
        type: string
      password:
        description: パスワード
        example: password
        minLength: 1
        type: string
    required:
    - id
    - password
    title: LoginPayload
    type: object
...
  SignupPayload:
    description: 新規登録時のメール送信
    example:
      email: yKicchanApp@gmail.com
    properties:
      email:
        description: メールアドレス
        example: yKicchanApp@gmail.com
        format: email
        type: string
    required:
    - email
    title: SignupPayload
    type: object
  Token:
    description: アクセストークン (default view)
    example:
      token: Bearer TokenString
    properties:
      token:
        description: アクセストークン
        example: Bearer TokenString
        type: string
    required:
    - token
    title: 'Mediatype identifier: application/vnd.token+json; view=default'
    type: object
...
  UserPayload:
    description: アカウント作成時に受け取る情報
    example:
      id: yKicchan
      mail: yKicchanApp@gmail.com
      name: きっちゃそ
      password: password
      tel: 090-1234-5678
    properties:
      id:
        description: ユーザーID
        example: yKicchan
        maxLength: 15
        minLength: 4
        type: string
      mail:
        description: メールアドレス
        example: yKicchanApp@gmail.com
        format: email
        type: string
      name:
        description: 名前
        example: きっちゃそ
        maxLength: 50
        minLength: 1
        type: string
      password:
        description: パスワード
        example: password
        minLength: 8
        type: string
      tel:
        default: ""
        description: 電話番号
        example: 090-1234-5678
        type: string
    required:
    - id
    - password
    - name
    - mail
    - tel
    title: UserPayload
    type: object
...
host: localhost:8888
info:
  contact:
    email: yKicchanApp@gmail.com
    name: yKicchan
  description: 卒業制作のアプリケーションで使用するAPIです
  license:
    name: Apache License 2.0
    url: https://github.com/yKicchan/EvelyApi/blob/master/LICENSE
  title: EvelyApi
  version: ""
paths:
  /api/develop/v1/auth/signin:
    post:
      description: ログイン
      operationId: auth#signin
      parameters:
      - description: 認証時に受け取るログイン情報
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/LoginPayload'
      produces:
      - application/vnd.token+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Token'
        "400":
          description: Bad Request
      schemes:
      - http
      summary: signin auth
      tags:
      - auth
  /api/develop/v1/auth/signup:
    post:
      description: 新規登録
      operationId: auth#signup
      parameters:
      - description: アカウント作成時に受け取る情報
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UserPayload'
      produces:
      - application/vnd.token+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Token'
        "400":
          description: Bad Request
      schemes:
      - http
      summary: signup auth
      tags:
      - auth
  /api/develop/v1/auth/signup/send_mail:
    post:
      description: 新規登録用のメール送信
      operationId: auth#send_mail
      parameters:
      - description: 新規登録時のメール送信
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/SignupPayload'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      schemes:
      - http
      summary: send_mail auth
      tags:
      - auth
  /api/develop/v1/auth/signup/verify_token:
    get:
      description: 新規登録時のトークンのチェック
      operationId: auth#verify_token
      parameters:
      - description: トークン
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/vnd.email+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Email'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      schemes:
      - http
      summary: verify_token auth
      tags:
      - auth
  /api/develop/v1/events:
    get:
      description: イベント複数取得
      operationId: events#list
      parameters:
      - default: ""
        description: キーワード
        in: query
        maxLength: 50
        name: keyword
        required: false
        type: string
      - default: 10
        description: 取得件数
        in: query
        maximum: 50
        minimum: 5
        name: limit
        required: true
        type: integer
      - default: 0
        description: 除外件数
        in: query
        minimum: 0
        name: offset
        required: true
        type: integer
      - description: ユーザーID
        in: query
        name: user_id
        required: false
        type: string
      produces:
      - application/vnd.event+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EventCollection'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      schemes:
      - http
      summary: list events
      tags:
      - events
    post:
      description: |-
        イベント作成

        Required security scopes:
          * `api:access`
      operationId: events#create
      parameters:
      - description: イベント作成・編集で受け取るイベント情報
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/EventPayload'
      produces:
      - application/vnd.event+json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Event'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: create events
      tags:
      - events
  /api/develop/v1/events/{user_id}:
    get:
      description: イベント複数取得
      operationId: events#list#1
      parameters:
      - default: ""
        description: キーワード
        in: query
        maxLength: 50
        name: keyword
        required: false
        type: string
      - default: 10
        description: 取得件数
        in: query
        maximum: 50
        minimum: 5
        name: limit
        required: true
        type: integer
      - default: 0
        description: 除外件数
        in: query
        minimum: 0
        name: offset
        required: true
        type: integer
      - description: ユーザーID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/vnd.event+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EventCollection'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      schemes:
      - http
      summary: list events
      tags:
      - events
  /api/develop/v1/events/{user_id}/{event_id}:
    delete:
      description: |-
        イベント削除

        Required security scopes:
          * `api:access`
      operationId: events#delete
      parameters:
      - description: イベントID
        in: path
        name: event_id
        pattern: ^[0-9]{6,8}-[0-9]+$
        required: true
        type: string
      - description: ユーザーID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: delete events
      tags:
      - events
    get:
      description: イベント情報取得
      operationId: events#show
      parameters:
      - description: イベントID
        in: path
        name: event_id
        pattern: ^[0-9]{6,8}-[0-9]+$
        required: true
        type: string
      - description: ユーザーID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/vnd.event+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Event'
        "404":
          description: Not Found
      schemes:
      - http
      summary: show events
      tags:
      - events
    put:
      description: |-
        イベント編集

        Required security scopes:
          * `api:access`
      operationId: events#modify
      parameters:
      - description: イベントID
        in: path
        name: event_id
        pattern: ^[0-9]{6,8}-[0-9]+$
        required: true
        type: string
      - description: ユーザーID
        in: path
        name: user_id
        required: true
        type: string
      - description: イベント作成・編集で受け取るイベント情報
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/EventPayload'
      produces:
      - application/vnd.event+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Event'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: modify events
      tags:
      - events
...
  /api/develop/v1/files/upload:
    post:
      description: |-
        ファイルアップロード

        Required security scopes:
          * `api:access`
      operationId: files#upload
      produces:
      - application/vnd.file_path+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/File_pathCollection'
        "400":
          description: Bad Request
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: upload files
      tags:
      - files
  /api/develop/v1/users/{user_id}:
    get:
      description: アカウント情報取得
      operationId: users#show
      parameters:
      - description: ユーザーID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/vnd.user+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      schemes:
      - http
      summary: show users
      tags:
      - users
...
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
...
schemes:
- http
securityDefinitions:
  jwt:
    description: |2-


      **Security Scopes**:
        * `api:access`: API access
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"